requirements.txt
alembic==1.7.1
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
dnspython==2.1.0
email-validator==1.1.3
flasgger==0.9.5
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
importlib-resources==5.2.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
passlib==1.7.4
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
